+++ packages/util-crypto/src/hmac/index.ts
@@ -1,0 +1,5 @@
+// Copyright 2017-2021 @polkadot/util-crypto authors & contributors
+// SPDX-License-Identifier: Apache-2.0
+
+export { hmacSha256 } from './sha256';
+export { hmacSha512 } from './sha512';