+++ packages/util-crypto/src/index.ts
@@ -13,6 +13,7 @@ export * from './blake2';
 export * from './crypto';
 export * from './ethereum';
 export * from './hd';
+export * from './hmac';
 export * from './keccak';
 export * from './key';
 export * from './mnemonic';